Adding a tip to run `mix hex.sponsor`
When a package that has a sponsor link is added or upgraded
